seperate triplet count
