+++ docbrown/db/src/algorithms.rs
@@ -31,3 +31,4 @@ pub mod directed_graph_density;
 pub mod local_clustering_coefficient;
 pub mod local_triangle_count;
 pub mod reciprocity;
+pub mod triplet_count;
